# src/main/java/mg/hei/heicorrectorapi/model/exception/BadRequestException.java
package mg.hei.heicorrectorapi.model.exception;

public class BadRequestException extends ApiException {
  public BadRequestException(String message) {
    super(ExceptionType.CLIENT_EXCEPTION, message);
  }
}